filmCards change sync
