+++ src/controllers/page.js
@@ -4,14 +4,14 @@ import SortComponent, {SortType} from "../components/sort.js";
 import MoreButtonComponent from "../components/more-button.js";
 import MovieController from "./movie.js";
 import {render, RenderPosition, remove} from "../utils/render.js";
-import {FILM_CARD_COUNT, SHOWING_FILMS_COUNT_ON_START, SHOWING_FILM_COUNT_BY_BUTTON} from "../const.js";
+import {FILM_CARD_COUNT, SHOWING_FILMS_COUNT_ON_START, SHOWING_FILM_COUNT_BY_BUTTON, EXTRA_FILM_CARD_COUNT} from "../const.js";
 
-let showedTaskControllers = [];
+const EXTRA_BLOCK_TITLE = [`Top rated`, `Most commented`];
 
 class PageController {
   constructor(containerClass) {
     this._films = [];
-    this._showedTaskControllers = [];
+    this._showedMovieControllers = [];
     this._containerClass = containerClass;
     this._sortComponent = new SortComponent();
     this._moreButtonComponent = new MoreButtonComponent();
@@ -58,21 +58,17 @@ class PageController {
     }
   }
 
-  render(films, minFilmsNum, maxFilmsNum, title = null) {
-    if (!title) {
-      render(this._siteMain, this._sortComponent, RenderPosition.BEFOREEND);
-      render(this._siteMain, new FilmsComponent(films), RenderPosition.BEFOREEND);
-    }
+  render(films, minFilmsNum, maxFilmsNum) {
+    this._films = films;
+    render(this._siteMain, this._sortComponent, RenderPosition.BEFOREEND);
+    render(this._siteMain, new FilmsComponent(films), RenderPosition.BEFOREEND);
 
     const siteFilmsSection = this._siteMain.querySelector(`.films`);
-    if (title) {
-      render(siteFilmsSection, new ExtraBlockComponent(title), RenderPosition.BEFOREEND);
-    }
     const siteFilmsListSection = this._siteMain.querySelector(this._containerClass);
     this._renderFilmCards(siteFilmsListSection, films, minFilmsNum, maxFilmsNum, this._onDataChange, this._onViewChange);
-    if (!title) {
-      this._renderShowMoreButton(siteFilmsSection, films);
-    }
+
+    this._renderShowMoreButton(siteFilmsSection, films);
+    this._renderExtraBlocks(films, siteFilmsSection);
 
     const siteFilmsListContainer = siteFilmsListSection.querySelector(`.films-list__container`);
     this._sortComponent.setSortTypeChangeHandler((sortType) => {
@@ -84,29 +80,42 @@ class PageController {
     });
   }
 
+  _renderExtraBlocks(filmCards, siteFilmsSection) {
+    EXTRA_BLOCK_TITLE.forEach((element) => {
+      const extraFilms = filmCards.slice().sort((prev, next) => {
+        return element === `Top rated` ? next.rating - prev.rating : next.comments.length - prev.comments.length;
+      }).slice(0, EXTRA_FILM_CARD_COUNT);
+
+      if (extraFilms.length) {
+        if ((element === `Top rated` && extraFilms[0].rating !== 0) || (element !== `Top rated` && extraFilms[0].comments !== 0)) {
+          this._containerClass = `.films-list--extra:last-child`;
+          render(siteFilmsSection, new ExtraBlockComponent(element), RenderPosition.BEFOREEND);
+          const siteFilmsListSection = this._siteMain.querySelector(this._containerClass);
+          this._renderFilmCards(siteFilmsListSection, extraFilms, 0, EXTRA_FILM_CARD_COUNT, this._onDataChange, this._onViewChange);
+        }
+      }
+    });
+  }
+
   _renderFilmCards(section, films, startNum, endNum, onDataChange, onViewChange) {
-    this._films = films;
     films.slice(startNum, endNum).forEach((film) => {
       const siteFilmsListContainer = section.querySelector(`.films-list__container`);
       const movieController = new MovieController(siteFilmsListContainer, onDataChange, onViewChange);
       movieController.render(film);
-      showedTaskControllers = showedTaskControllers.concat(movieController);
+      this._showedMovieControllers = this._showedMovieControllers.concat(movieController);
     });
   }
 
   _onDataChange(movieController, oldData, newData) {
-    const index = this._films.findIndex((it) => it === oldData);
-
-    if (index === -1) {
-      return;
-    }
-
-    this._films = [].concat(this._films.slice(0, index), newData, this._films.slice(index + 1));
-    movieController.render(this._films[index]);
+    this._showedMovieControllers.forEach((it) => {
+      if (it._filmCardComponent._filmCard === oldData) {
+        it.render(newData);
+      }
+    });
   }
 
   _onViewChange() {
-    showedTaskControllers.forEach((it) => it._setDefaultView());
+    this._showedMovieControllers.forEach((it) => it._setDefaultView());
   }
 }
